+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 * ``.values()`` and ``.values_list()`` now default to all fields if none are specified.
 * ``generate_schema()`` now generates well-formatted DDL SQL statements.
 * Fixed state leak between databas drivers which could cause incorrect DDL generation.
+* Added ``TruncationTestCase`` testing class that truncates tables to allow faster testing of transactions.
 
 0.12.7 (retracted)
 ------------------